Heatpump and buffer dynamics
Households can now heat their house with a heatpump and buffer. The behavior is defined in the heatwithheatpumpandbuffer function in households. Within the J_EAStorageHeat there are were some changes required: new boolean variable & and the maxTemp en setTemp variables are used in a different way than in the household heat model (@Gillis, pls check, maybe better if buffers are their on class)