Fix to contract loading logic in connectionOwner resulting in conflicts
Also, for connectionOwners now all financialTransactions are calculated in anylogic, instead of the cost-benefit module. This means that the contractData output will only have non-zero values in the financialTransactionVolume_eur variable, and the energyTransactionVolume_kWh will be 0 for all contracts.
This functionality can be rolled back to the previous implementation by setting the "B anylogic costs for all contracts" input to False.